+++ nixio/feature.py
@@ -5,9 +5,9 @@
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted under the terms of the BSD License. See
 # LICENSE file in the root of the Project.
-from ..entity import Entity
-from ..data_array import DataArray
-from .. import util
+from .entity import Entity
+from .data_array import DataArray
+from . import util
 
 
 class Feature(Entity):
