+++ nixio/pycore/__init__.py
@@ -1,4 +1,2 @@
 from .property import Property
 from .section import Section
-from .dimensions import RangeDimension, SetDimension, SampledDimension
-from .feature import Feature
